Wrote test for projection code and refactored and fixed the projection  module.
Projection
 -fixed append
 -fixed rotation matrix. It was mapping from earth to drone rather than
    drone to earth added transpose that seemed to fix it.
 -broke find target cords into _ned_to_geodic and _project_to_ground.

 Projection_tests
-Not all of theses work yet wont be pushing to master till they do.

Revert "Added the commented and fixed projection module"

This reverts commit d30264ad0de4e9a962f4023e63a4172b84e16274.

# Conflicts:
#	src/task1/projection-module/projection.py
